Add Footer and About draft components
