--- src/Frontend/App.1.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import './css/App.css';
-import AboutUs from './AboutUs';
-import Footer from './Footer';
-
-
-function App() { 
-  return (
-    <div className="App">
-      <button type="button">Accessibility Live</button>
-      
-    </div>
-  );
-}
-
-export default App;
