+++ src/Frontend/component/About/AboutUs.jsx
@@ -1,11 +1,13 @@
+import React from 'react';
 // import AboutUs1 from './Images/AboutUs.jpeg';
 
 export default function AboutUs() {
   return [
-    'about us', 'About Us', `At Team Charlie, we want to facilitate people with a disability to travel on public transport with ease. Accessibility Live is our signature application for improving the transport experience of people with a disability when using transport in and around the city of London.
+    <h1>About Us</h1>,
+    <h2>At Team Charlie, we want to facilitate people with a disability to travel on public transport with ease. Accessibility Live is our signature application for improving the transport experience of people with a disability, when using transport in and around the city of London.
     Accessibility Live presents key information that will enhance the experience of people with a disability. The information is based on knowing what stations have a working lift and when a particular station is busy. 
     
     Team Charlie's mission is for everybody to have a pleasant experience using public transport. At Team Charlie we believe that using current transport data to inform passengers, will allow them to make informed decisions about the best route of transportation for their needs.
-    We want our Accessibility Live application to be the leader in helping as many people as possible.`
+    We want our Accessibility Live application to be the leader in helping as many people as possible.</h2>
 ];
 }
