+++ src/Frontend/component/Footer/FooterCSS.css
@@ -1,0 +1,55 @@
+
+.container{
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(50px, 1fr));
+  flex-direction: row;
+  justify-content: center;
+  position: absolute;
+}
+
+.contacts {
+  flex-direction: row;
+  border: 1px solid rgb(24, 23, 23);
+  position: relative;  
+  overflow: hidden;
+}
+
+.img {
+  width: 10%;
+  height: 10%;
+  margin: 0%;
+}
+
+.alivelogo {
+  width: calc(3% - 2rem);
+  grid-template-columns: repeat(auto-fit, minmax(1px, 1fr));
+  justify-content: left;
+}
+
+.copyright {
+  width: calc(100% - 2rem);
+  justify-content: center;
+}
+
+
+
+/* @keyframes Footer-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+} */
+
+/* .Twitter-logo
+<div>Icons made by <a href="https://www.flaticon.com/authors/bogdan-rosu" title="Bogdan Rosu">Bogdan Rosu</a> from <a href="https://www.flaticon.com/" 			    title="Flaticon">www.flaticon.com</a> is licensed by <a href="http://creativecommons.org/licenses/by/3.0/" 			    title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div>
+
+.Facebook-logo
+<div>Icons made by <a href="https://www.freepik.com/" title="Freepik">Freepik</a> from <a href="https://www.flaticon.com/" 			    title="Flaticon">www.flaticon.com</a> is licensed by <a href="http://creativecommons.org/licenses/by/3.0/" 			    title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div>
+
+.Gmail-logo
+<div>Icons made by <a href="https://www.flaticon.com/authors/icomoon" title="Icomoon">Icomoon</a> from <a href="https://www.flaticon.com/" 			    title="Flaticon">www.flaticon.com</a> is licensed by <a href="http://creativecommons.org/licenses/by/3.0/" 			    title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div>
+
+.Github-logo
+<div>Icons made by <a href="https://www.flaticon.com/authors/roundicons" title="Roundicons">Roundicons</a> from <a href="https://www.flaticon.com/" 			    title="Flaticon">www.flaticon.com</a> is licensed by <a href="http://creativecommons.org/licenses/by/3.0/" 			    title="Creative Commons BY 3.0" target="_blank">CC 3.0 BY</a></div> */
